Update header.php
issue #118 added "listen" and "donate" mobile buttons into header onto line 160, also i moved the <br/> from line 158 up a couple lines so its inside div.example just above it so when we switch to mobile there's not an empty space above the navigation. 

but i noticed that was fixed recently, possibly removed, so maybe the <br> is all that is different from production site.